% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htm_to_md_img_lnk.R
\name{htm_to_md_img_lnk}
\alias{htm_to_md_img_lnk}
\title{HTML to Markdown Image Link}
\usage{
htm_to_md_img_lnk(x)
}
\arguments{
\item{x}{html object to convert to image}
}
\description{
HTML to Markdown Image Link
}
\examples{
\dontrun{
## copy to clipboard
library(table1)

reprexex::htm_to_md_img_lnk(table1(~mpg, data = mtcars))
## end copy to clipboard

reprex::reprex()
}
}
