% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_call.R
\name{remove_function_call_from_clip}
\alias{remove_function_call_from_clip}
\title{Remove Function Call from Clip}
\usage{
remove_function_call_from_clip(call = "as_img")
}
\arguments{
\item{call}{function call to remove}
}
\value{
invisibly returns clip code with function call removed
}
\description{
Remove the first function call from code on the system clipboard. Will
remove the call and both open and trailing parentheses but will leave inner
code if any.
}
\examples{
if (interactive()) {
  c(
    "library(table1)",
    "as_img(table1(~ mpg, data = mtcars))"
  ) |>
    clipr::write_clip()

  remove_function_call_from_clip()

  clipr::read_clip()
}
}
