% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_call.R
\name{remove_function_call_from_code}
\alias{remove_function_call_from_code}
\title{Remove Function Call from Code}
\usage{
remove_function_call_from_code(code, call)
}
\arguments{
\item{code}{code from which to remove function call}

\item{call}{function call to remove}
}
\value{
code with function call removed
}
\description{
Remove the first function call from code. Will remove the call and both
open and trailing parentheses but will leave inner code if any.
}
\examples{
c(
  "library(table1)",
  "as_img(table1(~ mpg, data = mtcars))"
) |>
  remove_function_call_from_code(call = "as_img")
}
