% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_call.R
\name{remove_library_call_from_clip}
\alias{remove_library_call_from_clip}
\title{Remove Library Call from Clip}
\usage{
remove_library_call_from_clip(lib = "reprexex")
}
\arguments{
\item{lib}{library for which to remove library call}
}
\value{
invisibly returns clip code with library call removed
}
\description{
Remove Library Call from Clip
}
\examples{
if (interactive()) {
  c(
    "library(reprexex)",
    "library(table1)",
    "as_img(table1(~ mpg, data = mtcars))"
  ) |>
    clipr::write_clip()

  remove_library_call_from_clip()

  clipr::read_clip()
}
}
