% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_call.R
\name{remove_library_call_from_code}
\alias{remove_library_call_from_code}
\title{Remove Library Call from Code}
\usage{
remove_library_call_from_code(code, lib)
}
\arguments{
\item{code}{code from which to remove library call}

\item{lib}{library for which to remove library call}
}
\value{
code with library call removed
}
\description{
Remove Library Call from Code
}
\examples{
c(
  "library(reprexex)",
  "library(table1)",
  "as_img(table1(~ mpg, data = mtcars))"
) |>
  remove_library_call_from_code("reprexex")
}
