% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reprexex.R
\name{reprexex}
\alias{reprexex}
\title{Render a reprex}
\usage{
reprexex(html_preview = opt(TRUE), ...)
}
\arguments{
\item{html_preview}{Logical. Whether to show rendered output in a viewer
(RStudio or browser). Always FALSE in a noninteractive session. Read more
about \code{\link[=opt]{opt()}}.}

\item{...}{additional arguments passed to \code{\link[reprex:reprex]{reprex::reprex()}}}
}
\value{
Character vector of rendered reprex, invisibly.
}
\description{
A post processor for \code{\link[reprex:reprex]{reprex::reprex()}}. First runs \code{\link[reprex:reprex]{reprex::reprex()}}. Then
Removes the \code{library(reprexex)} and \code{\link[=as_img]{as_img()}} calls from the clipboard
contents. Finally re-renders the html preview if requested.
}
\examples{
library(clipr)

if (clipr_available()) {
  c(
    "library(reprexex)",
    "library(table1)",
    "",
    "as_img(table1(~ mpg, data = mtcars))"
  ) |>
    write_clip()

  reprexex()
}
}
\seealso{
\code{\link[reprex:reprex]{reprex::reprex()}}
}
